strlcpy changes
darcs-hash:20070111171400-7aa3a-4c05d23c575a5dc59ad870a03dab0098e05321c2

+++ src/strlcpy.c
@@ -2,7 +2,7 @@
 
 #include "strlcpy.h"
 
-int strlcpy(char *dst, const char *src, unsigned int dstsize)
+size_t strlcpy(char *dst, const char *src, size_t dstsize)
 {
 	if (strlen(src) < dstsize) {
 		strcpy(dst,src);
